+++ fugerit-micronaut4-template/src/main/java/org/fugerit/java/template/Application.java
@@ -1,0 +1,18 @@
+package org.fugerit.java.template;
+
+import io.micronaut.runtime.Micronaut;
+import io.swagger.v3.oas.annotations.*;
+import io.swagger.v3.oas.annotations.info.*;
+
+@OpenAPIDefinition(
+    info = @Info(
+            title = "fugerit-micronaut4-template",
+            version = "0.0"
+    )
+)
+public class Application {
+
+    public static void main(String[] args) {
+        Micronaut.run(Application.class, args);
+    }
+}