+++ fugerit-micronaut4-template/src/main/java/org/fugerit/java/template/FugeritMicronaut4TemplateController.java
@@ -1,0 +1,12 @@
+package org.fugerit.java.template;
+
+import io.micronaut.http.annotation.*;
+
+@Controller("/fugerit-micronaut4-template")
+public class FugeritMicronaut4TemplateController {
+
+    @Get(uri="/", produces="text/plain")
+    public String index() {
+        return "Example Response";
+    }
+}